pyproject.toml
[tool.poetry]
name = "terraform-operator"
version = "0.1.0"
description = "Manages datacenter infrastructure through Terraform directly from Kubernetes."
authors = ["נυαη נυαηѕση <john@dewey.ws>"]

[tool.poetry.dependencies]
pykube-ng = "^21.3.0"
python = "^3.8"
kopf = "^1.32.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
coverage = {extras = ["toml"], version = "*"}
flake8 = "*"
hacking = "*"
isort = "*"
pyfakefs = "*"
pytest = "^6.0.1"
pytest-clarity = "*"
pytest-cov = "*"
pytest-mock = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
